refs #41563 main pcb done, todo: leds pcb multiplexor

+++ src/lib/esp8266/esp8266.cpp
@@ -27,7 +27,7 @@ char* Esp8266::strstr(char *haystack, const char *needle) {
 }
 
 char* Esp8266::strcat(char *dest, const char *src) {
-    size_t i,j;
+    uint32_t i,j;
     for (i = 0; dest[i] != '\0'; i++)
         ;
     for (j = 0; src[j] != '\0'; j++)
